+++ release-notes.md
@@ -132,6 +132,8 @@ python-bitcointx is now much easier. See for example
     p2wsh, and on integer values on witness stack
   - `de()serialize` and `stream_(de)serialize` now always work with `**kwargs`
     instead of additional params being passet as dict
+  - VectorSerializer `stream_(de)serialize` method signatures changed
+    to match base class
  
 * Breaking public API changes:
     - `CBitcoinAddress(<testnet_or_regtest_address>)` won't work: you will need to use `CCoinAddress` (universal, the class of returned instance depends on current chain params), or `CBitcoinTestnetAddress`/`CBitcoinRegtest` address directly. `CBitcoinAddress` is used only for Bitcoin mainnet addresses.